drivers/storage/sdhci: Use SDR25 presets for DDR50
The DDR50 and SDR50 presets currently use driver strength A. This
doesn't work on the zork platform. For now work around the issue by
using a different preset.

This is a nop for all other platforms. Zork is currently the only
platform using SDHCI presets. I will revert this CL once
http://crrev.com/i/3251807 lands. I'm waiting to land that CL until I
understand exactly which knobs we need. The initial implementation
assumed the same driver strength could be used across all timings, but
this turned out to be wrong.

BUG=b:159823235
BRANCH=zork
TEST=Boot ezkinil in HS400 and DDR52 mode.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Id48f36ae633562fc4a0bccc50c69b90b7dfd02e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2412877
Reviewed-by: Aaron Durbin <adurbin@google.com>