drivers/storage/mmc: Add timings for MMC_LEGACY and SD_DS
LEGACY currently defines timings for both MMC and SD. This change
breaks them out so it's clear which speed should be used. My goal is to
rename MMC_LEGACY into MMC_INIT. This will define the initial 400Khz
timing that is defined by the eMMC and SD specs.

BUG=b:159823235
BRANCH=zork
TEST=none

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I97e3947b009f39b0b30387280d736224ba38755c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2415389
Reviewed-by: Aaron Durbin <adurbin@google.com>