drivers/storage/sdhci: Add support for SD_DS and MMC_LEGACY
Now that we have a flag indicating 3.3V or 1.8V we can use the 3.3V
presets and the UHS-I presets when using 1.8V.

I also make the High Speed Enabled logic clearer. The SDHCI spec says
the flag is a "don't care" when 1.8V signaling is enabled. This is a
change in existing behavior.

BUG=b:159823235
BRANCH=zork
TEST=Boot ezkinil in HS and HS400 mode.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Icb0519f20867d50ca5f3825a846af1bec0ef1cc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2415390
Reviewed-by: Aaron Durbin <adurbin@google.com>