drivers/storage/{mmc,sdhci}: Rename MMC_TIMING_LEGACY to INITIALIZATION
Now that MMC_TIMING_LEGACY no longer has any callers, we can repurpose
it to specify the initialization timings. Since there are no callers
this change is a NOP.

The goal is to make mmc_setup_media call
`mmc_set_timing(INITIALIZATION)` instead of `mmc_set_clock(ctrlr, 1)`.
This way the platform can specify it's maximum open-drain speed. Dalboz
is missing a pull-up resistor on the CMD line so it can't run at 400KHz.
I still plan on implementing a fallback to a slower clock in depthcharge
if there are communication errors, but if the preferred clock could be
set by the platform it would avoid depthcharge having to guess.

BUG=b:159823235
BRANCH=zork
TEST=Boot ezkinil in HS and HS400 mode. Verified SD cards still get
discovered.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I294ff439d424a1bbe492dddb5d1f69f4f12b5bd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2415393
Reviewed-by: Aaron Durbin <adurbin@google.com>