spi: sc7180: Fix memory leak
They say memory leaks basically don't matter in a bootloader, which is
more or less true. But if you're leaking memory on every single SPI
chip select, it might still end up being enough to cause a problem in
the end. This patch fixes the qspi_sc7180 driver to not do that.

BRANCH=None
BUG=None
TEST=Booted Trogdor.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Id8a5e4a5e21378a8f953ab379540ba3eaecb41d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2415529
Reviewed-by: Stephen Boyd <swboyd@chromium.org>