trogdor/bubs: defconfig: Bump heap size to 24MB
A recent libpayload patch added framebuffer double-buffering, which
means we need to find space for another full framebuffer in the heap
(which comes out to about 8MB for 1080p). Together with the existing
requirements (e.g. another 8MB to back the flash for reading), our
current 16MB heap is no longer enough. Bump it to 24MB.

BRANCH=None
BUG=None
TEST=Booted Lazor.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I4960bcf466a7db07df30b1b93e550b65dabb8da3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2416834
Reviewed-by: Stephen Boyd <swboyd@chromium.org>