drivers/storage/sdhci: Disable clock before changing timings
The SDHCI spec says that when using presets, we need to disable the
clock before setting the high speed bit or changing the UHS-I timings.

This CL moves the code in set_ios around so that the clock is disabled
before setting the timing, and then enabled after the timings have been
set.

BUG=b:159823235, b:158766134
BRANCH=zork
TEST=Boot ezkinil to OS in HS400. Also trace registers and verify timing
is set before enabling clock.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Id9a4f2562d5abdaa8e4ddabd2c5928fa46f6fa76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2432991
Reviewed-by: Eric Peers <epeers@google.com>