drivers/storage/mmc: Add presets_enabled flag
We need a flag in the mmc layer to tell it that the underlying driver is
using presets. This way the mmc layer doesn't update the clock when
setting timings. The mmc layer still has the option of using
mmc_set_clock to explicitly override the clock set by any presets if
necessary.

BUG=b:159823235, b:158766134
BRANCH=zork
TEST=Build test

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I0af6288069479c7041cfa68cb770676db43972cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2432995
Reviewed-by: Eric Peers <epeers@google.com>