drivers/storage/sdhci: Remove zork preset hacks
Now that presets are set correctly we can remove all the TODO items.
I also refactored the switch statement a little bit to group all the MMC
timings together.

This change is a no-op because the driver strengths of the new 
preset registers are identical to the previous preset registers.

BUG=b:159823235
BRANCH=zork
TEST=Boot ezkinil to HS400.

Cq-Depend: chrome-internal:3251807, chromium:2436141
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I7933690a934cbca34d845dab05d07b1d74f391b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2432986
Reviewed-by: Eric Peers <epeers@google.com>