drivers/storage/sdhci: Use preset frequency if clock hasn't changed
This change query the preset registers and set the clock when only
the timing changes. This change is currently a no-op because the mmc
layer always sets the clock. The follow up patches make the mmc layer
stop setting the clock when presets are enabled.

Why not just set the PresetEn bit? I originally started a series where
I tried this approach, but it gets more complicated than just setting
the presets manually. Since there is no bus timing for initialization,
the SDHCI spec says the driver must manually query the preset register
and set the values. Once we have this logic, it's easier to just set the
presets for all timings, rather than special case initialization and
modify the code to not update the preset values for the other timings.

BUG=b:159823235, b:158766134
BRANCH=zork
TEST=Boot ezkinil to OS

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I36019dab89aeebdd3d49fa4e8edcd516db958e1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2432994
Reviewed-by: Eric Peers <epeers@google.com>