storage: sdhci: Avoid running sdhci_pre_init() twice
sdhci_pre_init() runs only once for fixed devices, but it runs on every
new insertion for removable devices. This is problematic on platforms
that have BROKEN_CLOCK_BASE set, because then the operation is not quite
idempotent: the platform is supposed to initialize host->clock_base in
MHz before the first sdhci_update(), but then sdhci_pre_init() converts
that value into Hz. When it is called a second time, the value is again
multiplied by one million, pushing it way out of range.

This patch applies a quick fix to the problem by preventing the function
from running more than once. It should otherwise be a no-op because it
only initialized values in the SdhciHost data structure. The only
hardware-interacting thing it does is reset the controller, which was
moved out of the function now.

In the long term, it would probably also be a good idea to prevent the
on-the-fly unit shift in host->clock_base and just have it provided in
Hz from the start.

BRANCH=None
BUG=b:170175862
TEST=None

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I830cc7c9f88d2ae219e15f79fe01029a71955f25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2454296
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2477413
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>