storage: sdhci: If SD cards fail to initialize, don't keep trying
When our SDHCI stack tries to initialize a removable card and fais, it
will just try again on the next sdhci_update() call. That is polled by
vboot in a pretty tight loop so this happens many times a second.
Generally, SD cards don't tend to be that flaky, so when a card cannot
be initialized for some reason, there's usually no need to keep trying.
This just tends to lead to a lot of log spam.

This patch makes sdhci_update() try each removable card up to 3 times,
and then if it still fails it won't try again until the card got
physically removed again.

BRANCH=None
BUG=b:170175862
TEST=None

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ia44e2bb960a9a954b20728a33d7e9bfecc53b795
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2459666
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2477414
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>