trogdor: Initialize the lpass_core_hm GDSC consistently
The lpass_core_hm GDSC has the retain FF bit set in the kernel but not
here. Not setting the bit here causes the clk registers inside the GDSC
to lose their state when the GDSC is disabled, meaning we spend more
time setting up the PLL and enabling clks in the kernel each time the
GDSC is turned off. Currently the kernel thinks the retain FF bit is set
at boot and turns off the GDSC after configuring the clks. This causes
the kernel confusion because the state of the clks are lost and
eventually the kernel PLL locks up.

Let's set the retain bit here and the waiting counts of the GDSC so that
it matches how the kernel handles this GDSC. The kernel implementation
should set the retain bit at device probe time to avoid this problem but
it's nice to be consistent and enable the GDSC the same way in
depthcharge and later on.

BUG=b:169292570
TEST=Boot on lazor, hear boot beep, see kernel doesn't print warnings
about stuck audio PLL
BRANCH=None

Change-Id: I635be444e0cf8f42813ccce2a9fa45cb1c069253
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2482330
Reviewed-by: Taniya Das <tdas@qualcomm.corp-partner.google.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2527286
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Reviewed-by: Martin Roth <martinroth@google.com>