WIP: ps8751: set MTP load done before flashing
This implements a vendor recomended change to the parade 87xx/88xx
TCPC firmware update procedure. In some cases, the TCPC could not be
recovered from an aborted firmware update. This change improves the
ability to recover a TCPC in a bad state.

BRANCH=none
BUG=b:168682309,b:148564462

Change-Id: Ieee80bd01ad8eee96e8750e5cb13b28d4219f615
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2568165
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>