ps8751: set MTP load done before flashing
This implements a vendor recomended change to the parade 87xx/88xx
TCPC firmware update procedure. In some cases, the TCPC could not be
recovered from an aborted firmware update. This change improves the
ability to recover a TCPC in a bad state.

BRANCH=trogdor
BUG=b:168682309,b:148564462
TEST=none

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: I1afc7acb8e9c96665bbe74694e9a06c214e19e06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2603400
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>