depthcharge/ui: Fix log initialization
In CL:2522156, fields 'lines_per_page' and 'chars_per_line' are added to
the ui_log_info struct. However, the initialization of these fields in
ui_log_init() is too early, so that they will be reset to zero in the
memset call.

Fix the initialization by keeping the original local variables, and
assigning them to the two fields of log after the memset call.

BUG=b:165187866
TEST=emerge-asurada depthcharge
TEST=Debug info and firmware log screens look good
BRANCH=none

Change-Id: I3622df0710ee2bf619ca78fc22b5c671bf55fe0e
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2554119
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2568505
Reviewed-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>