symbols: Declare the symbols as array type and unify with coreboot type
We would like to include the related header files in coreboot in
CL:2562866, then we have to align the symbol definitions.

The type of symbols in coreboot is "array of char". This design could
prevent error without an address-of operator in front of symbols.
However, the minus operator is not defined for "the pointer of array
type", i.e. "&ary_a - &ary_b" is invalid while "T ary_a[], ary_b[];"
This CL migrates that form to the appropriate one.

BUG=b:171858277
BRANCH=trogdor
TEST=build code with all archs
	emerge-zork depthcharge # (x86)
	emerge-nyan depthcharge # (arm)
	emerge-asurada depthcharge # (arm64)

Signed-off-by: Meng-Huan Yu <menghuan@chromium.org>
Change-Id: I6884bf4eaebdffe33b205b017ec866569a73add2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2577108
Reviewed-by: Julius Werner <jwerner@chromium.org>