sound: rt1015p: Support pre-init to reduce delay in calibration
The calibration time on rt1015p is too long (500ms) and would cause
the duration between 2 vboot security beeps to short (which should
happen on 20.0s and 20.5s after developer screen is shown).

To fix that, we want to pre-init I2S and audio amplifier when we believe
the speaker will be used (e.g., recovery mode or developer mode - which
both have screen initialized).

BUG=b:174534548
TEST=None
BRANCH=None

Change-Id: Ia553b17c326cf0299ed7469060a2bcac31e082c1
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2591831