storage: sdhci: Don't avoid running sdhci_pre_init() twice
In commit 2b1a609342a7 ("storage: sdhci: Avoid running
sdhci_pre_init() twice") (AKA CL:2454296) you can see that we found
that sdhci_pre_init() wasn't idempotent if
SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN was set. In that commit, we worked
around this problem by only calling it once.

While that seems like a good idea at first glance, there's a problem.
Specifically this means that host->attach() isn't called the 2nd time.
The one user of "host->attach()" is sdhci_msm.c and at the end of the
attach you can see this wonderful comment:
  * Set SD power off, otherwise reset will result in pwr irq.
  * And without setting bus off status, reset would fail.

When we avoid running the ->attach() the 2nd time we still do
sdhci_reset(host, SDHCI_RESET_ALL). Just as described in the comment,
that fails and the controller is basically in a super bad state.

Let's do a much less hacky solution for
SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN. We'll just to the multiplication
in the case that SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN wasn't set. If
SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN was set we'll assume it's in MHz.
We'll adjust the two users of SDHCI_QUIRK_CAP_CLOCK_BASE_BROKEN.

After this fix the retries are much faster.  However, it should be
noted, that the SD card still doesn't work if it wasn't plugged in at
boot.  That problem is not new and will need to be addressed by a
future patch.

BRANCH=trogdor
BUG=b:177688375
TEST=The 2nd attempt at enumerating the card no longer hangs

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I3e14d2d5b36c86b8ef32fd373465d881e43c8317
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2648944
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>