src/netboot: Run vboot initialization function as part of netboot
Netboot entry currently does not run vboot initialization functions.
One of the vboot init functions include registering a function to
override the lid close event. Without this function when the lid is
closed during the factory flow, the device shuts down. That behavior in
turn leads to unwanted side-effects during the factory flow.

BUG=b:178679255,b:178717179
BRANCH=None
TEST=emerge-dedede depthcharge. Build and boot the netboot BIOS to
depthcharge and make it wait for link. Ensured that the shutdown during
the lid close event is disabled as part of netboot entry.
  Starting netboot on Drawcia...
  WARNING: can't convert coreboot GPIOs, 'lid' won't be resampled at runtime!
  WARNING: can't convert coreboot GPIOs, 'power' won't be resampled at runtime!
  display: display_init called but not implemented.
  display: backlight_update called but not implemented.
  EC: Disabling lid close event due to GBB override
  ec_init: CrosEC protocol v3 supported (256, 256)
  EC: disabled lid close event
  Waiting for link
Close and open the lid to ensure that the device does not shutdown.
  21-01-28 13:52:32.104 [47.349592 lid close]
  21-01-28 13:52:32.104 [47.350008 KB disable_scanning_mask changed: 0x00000001]
  21-01-28 13:52:32.111 [47.350602 KB Clear Buffer]
  21-01-28 13:52:32.114 [47.351845 SW 0x04]
  21-01-28 13:52:35.520 [50.783318 lid open]
  21-01-28 13:52:35.520 [50.783735 KB disable_scanning_mask changed: 0x00000000]
  21-01-28 13:52:35.527 [50.786979 SW 0x05]
Ensure that the link is established when a type-C to ethernet adapter is plugged in and
an IP address is assigned.
  Waiting for link
  R8152: Initializing
  Version 5 (ocp_data = 5c20)
  R8152: Done initializing
  Adding net device
  done.
  MAC: 9c:eb:e8:87:f9:b0
  Sending DHCP discover... done.
  Waiting for reply... done.
  Sending DHCP request... done.
  Waiting for reply... done.
  My ip is 192.168.0.109
  The DHCP server ip is 192.168.0.1
  TFTP server IP supplied by DHCP server: 0.0.0.0

Change-Id: Id90c327da576c816a7a41bb5e2606586a007fb2b
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2657775
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@chromium.org>
(cherry picked from commit a86dc37d95b2f0a8fcb0006d0cff4efd219715c4)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2683879
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Tested-by: Nick Vaccaro <nvaccaro@google.com>