herobrine: Add support for dma coherency
If DMA coherency is not supported, we need to invalidate
cache for read operation and clean the cache for write
operation before posting the command to submission queue.
This will ensure the CPU read the new contents from memory
after nvme controller completes the dma operation for read
and also data from the cache be updated in memory and be
visible to the DMA transfer.

Signed-off-by: Nitin Rawat <nitirawa@codeaurora.org>
Signed-off-by: Shaik Sajida Bhanu <sbhanu@codeaurora.org>
Change-Id: I82cb452ac3acb6338fa995bf4f12fb5ed294f52f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2835160
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>