treewide: Rename EC_SOFTWARE_SYNC to EC_VBOOT_SUPPORT
After coreboot added support for EC software sync, there was some
confusion over whether when a board selects coreboot to perform the
sync, is it still required to select EC_SOFTWARE_SYNC in the
corresponding depthcharge target as well. The answer turns out to be
'yes', but for a surprising reason: for recovery mode, vboot needs to
know if user input from the the EC can be trusted or not. Some boards
that had coreboot performing EC sync had removed the option from the
corresponding depthcharge target and thus introduced a security bug into
RO FW.

This patch intends to make it easier to do the right thing, by:
1) Renaming EC_SOFTWARE_SYNC to EC_VBOOT_SUPPORT. This should make it
   more clear that the option needs to be selected for Vboot support,
   regardless of whether software sync is performed by coreboot or
   depthcharge (in the end, it's OK to enable in both, because vboot
   includes its own logic to not perform EC software sync more than once
   per boot).
2) Move ec_util code back into ec.c

If a mainboard still does not want EC software sync to be performed, the
VB2_GBB_FLAG_DISABLE_EC_SOFTWARE_SYNC flag can be enabled in
coreboot. If the mainboard selects CONFIG_DRIVERS_CROS_EC but not
CROS_EC_VBOOT_SUPPORT, then an error about an undefined reference to
vb2ex_ec_battery_cutoff will be seen.

NB This patch was created via find/sed (defconfigs) and manual
inspection.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I0de0e3199ef949ef98fc8f88409495a243b5b5a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2802491
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>