drivers/ec/ps8751: Implement app updater for ps8815
This implements the ps8815 "application" firmware updater.  Recent
ps8815 firmware support the notion of a "base" firmware and an
"application" firmware. The base firmware is typically the factory or
pre-production image. To update the TCPC firmware in-field, only the
application region and associated header need to be programmed.  This
means there will always be a known-good base firmware image available
for recovery.

The firmware updater recognizes the type of the firmware image bundled
in CBFS (BIOS) and programs the appropriate flash region.

BRANCH=none
BUG=b:186189039,b:176602433
TEST=updated ps8815 firmware on brya with base firmware as well as
	application firmware. verified FW version changes as expected.

Change-Id: Ia78c83292f063e22ce1fcdb212e38c431cd6545f
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2858804
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>