zbi: use kernel command line from ZBI for substitution.
This adds a hook, zbi_fill_boot_info(), which is called by
multiboot_fill_boot_info().

This allows us to use the kernel command line included in the
ZBI (if one is present) to perform substitutions on, rather
than always having an empty command line.

Since the kernel command line item in the ZBI is exactly the size of the
string, we perform the replacement by marking the hardcoded entry as
'ZBI_TYPE_DISCARD', so that the kernel will ignore it, and adding a new,
larger entry with type 'ZBI_TYPE_CMDLINE' that contains the substituted
command line string.

Test: add %U to ZBI_TYPE_CMDLINE item in fuchsia.zbi, and verify it's
      correctly modified by depthcharge.
Bug: fuchsia:74800
Change-Id: Ia866df0b74b471ecd208e64c5ee26ceb47f5bb38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2858178
Commit-Queue: Simon Shields <simonshields@google.com>
Tested-by: Simon Shields <simonshields@google.com>
Reviewed-by: Shelley Chen <shchen@chromium.org>