fastboot: add support for slots.
This CL adds support for slots in fastboot. Since a device's partition
table is fairly flexible, we can't rely on a particular partition label
being 'a' or 'b'. Instead, each kernel partition (with the CrOS kernel
GUID) is treated as a separate 'slot', and the client can run `fastboot
oem get-kernels` to view the mapping of slot -> partition.

Bug: fuchsia:76065
Test: fastboot oem get-kernels
Test: fastboot set_active a && fastboot oem get-kernels && fastboot
      set_active b && fastboot oem get-kernels shows changed slot.
Test: fastboot getvar slot-count
Change-Id: Iceb8e83e55ef76509da829e23ae17b65e82c9225
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2873314
Commit-Queue: Simon Shields <simonshields@google.com>
Tested-by: Simon Shields <simonshields@google.com>
Reviewed-by: Shelley Chen <shchen@chromium.org>