vboot/Kconfig: deprecate config DIAGNOSTIC_UI
Config DIAGNOSTIC_UI is passed to vboot to enable either diag_payload
or minidiag.

For diag_payload:
It is for a diagnostic tool which based on legacy UI and deprecated.

For minidiag:
Mini-diag is set default enabled for newer devices, and we could still
use (vboot) kernel secdata in runtime if we decide to disable it on
certain platforms.

In conclusion, deprecating DIAGNOSTIC_UI will not affect newer
devices.

BUG=b:190796342, b:181931817
BRANCH=none
TEST=emerge-volteer depthcharge

Cq-Depend: chromium:3006114, chromium:3005643
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: I7690666f6456c1bf85e35937acf089e72fbf270f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3004223
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>