board/guybrush: Replace EC_IN_RW flag instead of install
In prior board versions, EC_IN_RW is not routed to AP and hence it was
installed. Starting board version 3, the signal is routed to GPIO_91 and
is going to be exported by coreboot. So for prior board versions,
replace the ECINRW flag to keep things simple at coreboot.

BUG=None
BRANCH=None
TEST=Build and boot to OS in Guybrush. Ensure that the system boots fine
in both recovery and normal modes.

Change-Id: I5ba04658dbcd640b47e2e55b0d3c442fd1b04967
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3043702
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@chromium.org>