board: Add Kconfig string option for variant name
Now that each build target will end up with its own FW_CONFIG, it may be
easier to manage variant creation if the common board setup is done in
board.c, and each variant handles their specific setup code in a
board-specific variant.c file or similar. This can be accomplished by
adding some logic to the Makefile to add only variant-specific files to
the build.

BUG=b:193527706
TEST=Test patchtrain

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I074f2b50c72ac668edcac25435305b80e0143a04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3049991
Reviewed-by: Furquan Shaikh <furquan@chromium.org>