drivers/soc: Add intel_remap_pcie_rp() function
When the FSP runs, it disables non-hotpluggable PCIe RPs that have no
downstream devices connected. Thus, there can be the case where a slot
no longer has a function #0, but other functions are present. In order
to remain PCI compliant (multi-function devices must have a function 0
present), one RP in the same slot is remapped to function 0
instead. This can be detected programmatically (coreboot does it),
therefore port this logic for depthcharge to use as well.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: Id601cab12f65f7a4235a0d95586bc6d34f3286d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3115371
Reviewed-by: Furquan Shaikh <furquan@chromium.org>