depthcharge/tests: Mock vboot_get_context
Some UI functions have references to vboot_get_context. In order
to obtain a valid pointer to ctx, we can mock the workbuf and use
vb2api_init to achieve.

Mock vboot_get_context, and add commonparams.h to export the workbuf.
Provide a variable reset_mock_workbuf in the header for
vboot_get_context to decide whether the workbuf needs to be reset.

BUG=b:172339016
TEST=sudo USE=abi_x86_32 emerge dev-util/cmocka && \
     make clean-unit-tests && make unit-tests
BRANCH=none

Signed-off-by: edisonhello <edisonhello@google.com>
Change-Id: I7779159c057af40571c3539359783cd34c73dcbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3127969
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>