vboot/callbacks: Deprecate vb2ex_ec_trusted()
coreboot will support checking the EC_IN_RW during verstage_main(),
which performs the equivalent tasks as vb2ex_ec_trusted().

In the previous CL, we remove the check of vb2ex_ec_trusted(), so the
whole implementation of this function among vboot_reference and
depthcharge could be deprecated and removed.

Also deprecate the FLAG_ECINRW since it is only used for determine if EC
is trusted.

BUG=b:181931817
BRANCH=none
TEST=CC=x86_64-pc-linux-gnu-clang;
     make clean && make runtests
TEST=emerge coreboot vboot_reference depthcharge

Cq-Depend: chromium:3139539
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: I52afb60b829b19dfa289ed1c0598b2d325011143
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3139956
Tested-by: roy johnson <roypjohnson76@gmail.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>