depthcharge/vboot: Internalize VbExKeyboardReadWithFlags
After centralizing the UI code in depthcharge, lots
of VbEx* and vb2ex_* functions are not used anymore in vboot.

VbExKeyboardReadWithFlags is one of them. Move the implementation to
vboot/keyboard.c, and rename it to ui_keyboard_read.

Since no other function is calling VbExKeyboardRead, rename the
function to keyboard_get_key, and make it a static function in
vboot/keyboard.c.

Move VbKeyFlags_t enum from vboot to depthcharge, and rename it to
ui_key_flag.

BUG=b:172339016
TEST=emerge-asurada depthcharge
TEST=make clean-unit-tests && make unit-tests
BRANCH=none

Cq-Depend: chromium:3143812
Signed-off-by: edisonhello <edisonhello@google.com>
Change-Id: I58dba04cc1a2fd29b985abd2f909ecb76afc8ea1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3142696
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>