depthcharge/ui: Implement scrollbar in log screen
Add a scrollbar at the right of log screen. The scrollbar has the same
width and corner radius as the one in the language manu.

Add another macro UI_SCROLLBAR_MIN_HEIGHT to limit the minimal height
of the scrollbar.

The position of the scrollbar in log screen will not be affected by
the language, since the log content is left aligned. Reverse
the position will overlap the text.

Since draw_language_select also has similar part in drawing the
scrollbar, add a ui_draw_scrollbar function that will calculate the
position and the height of the scrollbar automatically, and use it in
both draw_language_select and draw_log_desc.

BUG=b:191434445
TEST=Hayato booted into developer mode
TEST=Hayato showed firmware log screen with scrollbar
TEST=Hayato showed debug info screen with scrollbar
TEST=make unit-tests
TEST=emerge-asurada depthcharge
TEST=make screenshot SCREEN=UI_SCREEN_DEBUG_INFO \
     LOG_FILE=logfile PAGE=10
TEST=make screenshot SCREEN=UI_SCREEN_DEBUG_INFO \
     LOG_FILE=logfile PAGE=40
TEST=Changing ui_get_locale_count return value
     in screenshot/archive.c and run
     make screenshot SCREEN=UI_SCREEN_LANGUAGE_SELECT ITEM=15
BRANCH=none

Signed-off-by: edisonhello <edisonhello@google.com>
Change-Id: I079b87aa2e53c3dbafd8e47e22fbfb12c949e969
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3144674
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>