ps8751: Fix hardware version check for PS8805/PS8751
The Parade TCPC family doesn't consistently set the chip revision
registers. Add the Parade recommended secondary checks of the hardware
version for the PS8805 and PS8755 chip types.

Note that the previous fix on CL:3132309 was not checking the correct
bits in the REG_REV (Page 0, offset 0x62) register.

BUG=b:199550987, b:193099851
BRANCH=none
TEST=Verify Parade PS8805-A3 firmware update on Herobrine

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I52309ab6d8aa0026bace66cfcd51bd4c770e2050
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3159164
Reviewed-by: Wai-Hong Tam <waihong@google.com>