i2c/dw: Enforce register alignment at compile time
Instead of hoping that "packed" does the right thing (the compiler isn't
forced to) and confusing the compiler because a packed access might be
unaligned (even though it isn't here because it's all uint32_t), just
check that the last offset turns out right.

BUG=none
BRANCH=none
TEST=`emerge-zork depthcharge` passes with both gcc8 and gcc11. Changing
the offset it tests for to 0xb5 breaks the build as expected.

Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Change-Id: If0c628629aba06daa2e56b90776de75611f4e934
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3181723
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>