ui: Process user input before screen action
The (simplified) execution order of the current UI loop is:

1. Display screen
2. Process screen action
3. Process keyboard input (for menu navigation)
4. Process global action

The problem is, after screen action is run, the screen may change. Then
the user input processing will be run under the new screen instead of
the original one shown to the user.

Therefore, swap the order of #2 and #3, i.e, process user input before
screen action in the UI loop.

BUG=none
TEST=make unit-tests
TEST=emerge-brya depthcharge
BRANCH=none

Change-Id: Ic06a2562f1ae1cf86dc306b4ac3fd833fe9689da
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3195523
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>