ui: Utilize get_file() for test screens
With get_file() added in CL:3257501, now we can use it for test screens.
This avoids having to modify focused_item and hidden_item_mask in each
iteration, and makes the boundary between UI logic and drawing clearer.
When drawing the test screen, the correct bitmap file will be selected
based on the state of the test.

BUG=none
TEST=make unit-tests
TEST=make screenshot \
     SCREEN=VB2_SCREEN_DIAGNOSTICS_STORAGE_TEST_SHORT LOG=Sample
TEST=emerge-geralt depthcharge
TEST=Ran quick memory test on Geralt and saw "Cancel" changed to "Back"
BRANCH=none

Change-Id: I22c8e042c318fcf5c2ad16bd7a8eab52a5068dab
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3257502
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>