vboot/ui: Deprecate vb2api_allow_recovery()
2lib used vb2_api_allow_recovery() to differentiate between manual and
non-manual recovery in 2kernel and UI related areas.

With introducing the ctx->boot_mode, we could decide if it is a manual
recovery or a broken screen (a.k.a non-manual recovery in the original
design) once in vb2api_fw_phase1 and use this boot mode instead for
further justifications.

Also modify the corresponding mocks in unit tests.

BUG=b:181931817
BRANCH=none
TEST=make unit-tests
TEST=emerge-brya depthcharge

Cq-Depend: chromium:3121926
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: I07b5ab8a5b2fa48c7a58194dbad6e00abfde9673
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3282875
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>