screenshot: Add log file and current page options
Add two options: LOG_FILE and PAGE options.

LOG_FILE=<filename> will load the content of filename, and display
as log content. If LOG option is also specified, LOG_FILE will be
ignored.

PAGE=<number> controls the current page of the log, which
will change the displaying content when generating the image.

For example:
  make screenshot SCREEN=UI_SCREEN_DEBUG_INFO \
  LOG_FILE=logfile PAGE=10

Change the function read_bmp to more general read_file, and add the
parameter open_mode to specify which mode to use, "r" or "rb", to open
the file.

BUG=none
TEST=make screenshot
TEST=make screenshot SCREEN=UI_SCREEN_DEBUG_INFO \
     LOG_FILE=logfile PAGE=10; # With some custom logfile
BRANCH=none

Signed-off-by: edisonhello <edisonhello@google.com>
Change-Id: I0649af937ab97a5aa761d234c280e4c23f1517cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3289390
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>