drivers/gpio: Add new GPIO splitter building block
Similar to the virtual "and"/"or"/"not" GPIOs we have to tie underlying
input GPIOs together, it would be useful to have a virtual splitter GPIO
that just ties two underlying output GPIOs together (by forwarding all
output values to both). Since this would often be used for regulator
power sequencing, make it so deassertions happen in reverse order (b
before a) than assertions (a before b), and also give it the ability to
delay a set amount of time between each.

BRANCH=trogdor
BUG=b:206922063
TEST=Booted CoachZ

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I430e1fff6f623af56d16daf8b354f796ba234caf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3387601
Reviewed-by: Douglas Anderson <dianders@chromium.org>